feat: add modular package boundaries (0.60.0)
