+++ packages/cli/package.json
@@ -1,0 +1,5 @@
+{
+  "name": "@wendkeep/cli",
+  "private": true,
+  "type": "module"
+}
