+++ packages/harness/package.json
@@ -1,0 +1,5 @@
+{
+  "name": "@wendkeep/harness",
+  "private": true,
+  "type": "module"
+}
