+++ packages/integrations/package.json
@@ -1,0 +1,5 @@
+{
+  "name": "@wendkeep/integrations",
+  "private": true,
+  "type": "module"
+}
