+++ packages/mcp/package.json
@@ -1,0 +1,5 @@
+{
+  "name": "@wendkeep/mcp",
+  "private": true,
+  "type": "module"
+}
