+++ packages/pi/package.json
@@ -1,0 +1,5 @@
+{
+  "name": "@wendkeep/pi",
+  "private": true,
+  "type": "module"
+}
