+++ packages/vault/package.json
@@ -1,0 +1,6 @@
+{
+  "name": "@wendkeep/vault",
+  "private": true,
+  "type": "module",
+  "exports": "./src/index.mjs"
+}
