+++ packages/vault/src/index.mjs
@@ -1,0 +1,2 @@
+export * from './project-vault.mjs';
+export * from './vault-path-safety.mjs';
